Send e-mails with a Sender: header for better DKIM support.
Some spam filters get caught up on Review Board's From: forging. The address
doesn't really match the source address, causing them to expect spam or
some other malicious e-mail.

We now use the Sender header (and the X-Sender variant) in the e-mails.
Servers can use this to find the true sender of the e-mail, instead of assuming
it's the value in From. By default, this is noreply@<domain>, but it can be
changed in the admin UI.

This should improve our DKIM support.

The tradeoff is that some e-mail clients (Outlook, for instance) will show
some annoying "on behalf of" text where the sender's e-mail address is
typically shown. We may want to add the ability to disable this on installs.

While annoying, this is the behavior of mailman as well. See
http://wiki.list.org/display/DOC/From+field+displayed+by+Microsoft+Outlook

Reviewed at http://reviews.reviewboard.org/r/2033/