Fix API calls which send emails.
Sending email for review requests and reviews wasn't passing in the local site
name, which meant that if non-rooted URLs were turned off, the call would error
out, and if they weren't, the URLs included in the messages would just be wrong.
This fixes that by adding local_site_name to the context so our processor can
use it.