Make DefaultReviewer aware of LocalSites.
DefaultReviewer wasn't tied to a LocalSite, making it impossible to see
which ones exist when not bound to a Repository or Group. It now stores the
LocalSite, and the DefaultReviewerForm performs validation to ensure that
the repositories, users and groups selected match the LocalSite in order to
prevent people or code from linking to objects on other LocalSites.

Reviewed at http://reviews.reviewboard.org/r/2040/