Add LocalSite validation for Group's list of users.
This introduces new validation for Group that ensures the users selected
are actually part of the LocalSite, if the Group is a part of one.

This makes use of the same user validation logic as DefaultReviewer. The
checks have been split out into a _validate_users function, which both
can access.

Unit tests have been added to ensure that validation works how we expect.

Reviewed at http://reviews.reviewboard.org/r/2041/