Fix validation issues with the new LocalSite checks in admin forms.
The LocalSite checks were assuming that local_site would be "cleaned" prior
to validating the fields that depend on them. That was clearly a bad idea.
Validation consisting of multiple fields should be done in clean(), so I've
moved the calls to there. Nothing changed with the actual validation logic.
The calls are just in one place now.