Fix validation issues with the Repository.
This fixes another validation issue where the repository was also processing
fields in the wrong order. We now validate users and groups in clean().

This fix uncovered another problem, which is that if no entries were available
(say, no invite-only groups), our validation checks would raise an exception,
as the field wouldn't be in cleaned_data. We now default to an empty list in
these cases.