Display local site names in the user group membership UI
This change fixes the "My Account" form to show all groups across local sites,
and prefixes them with the local site name (if applicable). This way people will
be able to join groups even before I finish up the change to the group
membership UI to put it in the datagrid.