Add first and last names to the sign-up form, turn off interruption.
It seems a little annoying to have to fill these in from the preferences form
after you've already signed up. In addition, we would interrupt users asking
them to join groups even if there were no groups.

I've turned off the first-time-setup screen and message entirely for now, and
intend to repurpose it to show a "what now" screen with some help and basic
instructions.