Fix the unit tests.
Apparently the django docs' suggestion of hooking up to User's post_save signal
in order to create the Profile wasn't actually a good idea. Since we already
wrap things in valid_prefs_required, and that implementation wasn't actually
good anymore anyway, we'll just repurpose it to create the profile object.