Return 404 if the local site name doesn't exist.
I noticed the other day that accessing the dashboard with a local site name that
didn't exist showed me a dashboard page. This was happening because we used
get_object_or_none for the local site in a lot of places, which meant that said
views were showing the root data instead of 404.