Provide a "Show SSH Public Key" option in the Add/Change Repository page.
Finding the SSH public key can be difficult on a standard Review Board
setup, and if the user has access only to the repository parts of the
admin UI, they can't get the public key.

This provides a "Show SSH Public Key" button in the top-right of the
Add/Change Repository page, which will toggle the display of a box containing
the public key.