Implement several fixes to rbssh.
rbssh had problems with some types of repositories. Mostly, they expected
certain command line options to be available that weren't. We now support
-q/--quiet, -s (for subsystem selection), and -V (print version).

In general, the rbssh code has been cleaned up to be a bit better structured
and less hacky. We have classes for Posix and Windows, rather than
individual functions. We also now tell OptionParser to leave behind any
unknown arguments so that we can pass them to the command to run on the
other server.

rbssh now supports a shell and sftp. If a valid key or password wasn't
available during authentication, we prompt for a new password. We do this
three times before giving up.