Fix communication problems with rbssh on Windows.
This fixes problems with rbssh on Windows. The process was hanging around
forever instead of properly shutting down. We now let stdin processing
drive in the main thread and then, upon closing, we trigger a shutdown
of the writer, which should close the other end and shut down the process.