Support taking a repository name when creating a review request.
Historically, we've supported taking a numeric repository ID or a
repository path as an argument, which works fine when dealing with
setups with only one or two possible paths. However, for Git and other
such types of repositories, you may have a dozen different paths, since
they may be SSH-backed with a username. This causes creation to fail,
even though each of these may really be the same repository.

This supports taking a repository name in place of an ID or path when
creating a new review request. The name would be the exact same name you'd
see in the New Review Request page.

This requires tools to know that they can pass a name. RBTools will soon be
updated for this.