Fix search index validation on the General Settings page.
The new search index validation broke for any user that didn't have a
valid, writable, existing search index file specified. This was a problem
for new and existing users.

Now we only validate access if the file exists, and we don't validate
existence, since it's perfectly fine for it not to exist yet. We also fall
back early if nothing was specified.

Even once providing valid information, we weren't saving it, so users
could never get passed this stage. We now make sure to return the value
so it saves in the database.