Add a user page and hover preview.
Currently when a user visits /users/<user name>, all of the review requests of
that user are shown.  This patch modifies this page slightly to include more
information.  This is a first step, and I've tried to make it as simple as
possible.  Further modification will probably be needed.

I have also added a hover preview.  When the cursor moves over top of a user
name a small preview of that user will be shown (see screenshot).  This feature
is only supported on the review request pages (not in the "submitters" page for
example).  This could be added later, but was omitted due to the possibility of
too many images appearing.

Reviewed at http://reviews.reviewboard.org/r/1953/