Fix screenshot drag-and-drop.
Screenshot drag-and-drop broke in 1.6 due to a change to force the
mimetype for responses. I added code to add "api_format=json" to the
data dictionary, which did interesting things when data was not a dictionary
but instead was a string. It therefore lost string's functions, which broke
jQuery's ajax functions.

Now screenshot drag-and-drop works once again.