Allow usernames with '@' in the admin backend
It is an usual situation for LDAP servers use the mail attribute
for authentication. If ReviewBoard is set up to authenticate using
an email attribute, login is successful and the user data is
correctly populated. Nevertheless, it is not possible to change the user
field as the username is considered incorrect.

This commit relaxes the restriction in the admin backend to allow for
the '@' character in usernames. There is no need to perform a similar
change in the frontend code as LDAP registrations do not follow the same path
as basic registrations.

Reviewed at http://reviews.reviewboard.org/r/2124/