Fix appending the "Press Tab to auto-complete" footer on autocomplete dropdowns.
Only the first autocomplete dropdown used would receive the "Press Tab to
auto-complete" footer. We would look for the ".ui-autocomplete-results" and
only append if we didn't see a footer in it. However, once a second autocomplete
dropdown was shown, due to our query, we'd show that there was at least one
footer, and skip adding a new one.

Now we use a query to find specifically those results divs without a footer,
and add to all the ones found.