Don't show the groups box if there are no groups to list.
When there are no review groups defined, we don't want to clutter the user
preferences page with empty boxes or meaningless instructions.

Testing done: Ran this with a new user when there were no groups in the db at
all, and ran it with a new user when there were groups in the db.
Fixes bug 1233.
Reviewed at http://reviews.reviewboard.org/r/2043/