Fix bug 1571
On a "New Review Request" page, the text fields for paths and filenames are too
short. By setting the widths of related table and classes width 100% to fit the
screen, and the size of the fileinput to be long enough under 800x600, they
become longer.

Testing done: It works well in firefox (800x600) and Chromium.
Untested in IE and safari as well as for SVN repository.
Fixes bug 1571.
Reviewed at http://reviews.reviewboard.org/r/2058/