Incoming group count is not update after updating the review request with new values for Groups and People.
What steps will reproduce the problem?
1. Create a review request and assign it to user (People) say User1
2. Update the review request and change the People value say User2
3. Login as User2.

What is the expected output? What do you see instead? To-me count in the
dashboard should be incremented, which is not happening.

The same condition applies to the Groups field too, where the Incoming Reviews
count is not updated. Though the decrement on the count for the other user/group
happens good.

Manual Verification done.

Fixes bug 2032
Reviewed at http://reviews.reviewboard.org/r/2203/