Don't error out when updating an existing changenum that exists multiple times.
The "New Review Request" form tries to update an existing review request when it
gets a change number in use error. Unfortunately, the query was just looking up
the existing review request based on the changenum, without also passing in the
repository. This means for deployments with several change number-using
repositories tried to do this, it might fail.

Fixes bug 2024.