Don't associate draft screenshots with review requests until published.
We've always immediately associated a screenshot with a review request,
which meant users could see it even if it was part of a draft. This was
inconsistent with the way that diffs and field changes worked. It also
meant that the entries in change descriptions weren't correct.

Now we keep it on the draft only until published. Everything has been
updated to work with this change. The web UI just simply works, since
it's already looking at the list on the draft when present, and the
web UI and model now work properly when a review request isn't directly
associated (it's accessed through the draft in that case).