Fix ordering of values in change descriptions.
Change descriptions would end up with values being in seemingly random orders.
Bug numbers, for example. This was due to converting to sets prematurely.
We were converting to sets in order to get the add/remove lists, but there
was no reason to using them in the old/new lists.