Add API for getting change descriptions.
This introduces new API for getting change descriptions on a review request.
It will return all public change descriptions, along with the text, timestamp,
and a list of fields changed. The list of fields changed is a more detailed,
useful list than what is stored internally. We expand it out, embedding
resources with links and providing a useful representation for the data.

Fixes bug #2026