Remove RSS/Atom feed support.
In the very early days, we had RSS/Atom feeds for review requests, users and
groups. These aren't really useful anymore, and nothing even exposes them,
so it's unlikely this will break anybody (if the feeds even work anymore
anyway).

The motivation for finally killing this code is that Django changed how
feeds work in 1.2, and officially removed backwards-compatibility in 1.4.
It's not worth it to update them.