Update the test suite to use the new test runner APIs.
Django 1.2 introduced a new test runner API, which is class-based instead
of a single function like the old one. In the in-development Django 1.4,
these are deprecated and will fail.

This moves the existing code into a new class-based runner. The logic hasn't
changed, with the exception that we can now call into the parent class for
most of the setup work for databases, saving us some code and imports.
Since we no longer have to import some of the test suite setup functions,
we can also remove our ugly exclude list.