Float banners
This update adds a class 'float-element' that can be used to float elements at
the top or bottom of the screen if they are scrolled off-screen.

If you don't want an element to float everywhere on the screen you can add the
'float-element-container' class to an element wrapping the 'float-element' and
it will only float while it is within the container.