Make the code for float banners more performant and self-contained.
This improves upon Mark's float banner code to be a bit more performant.
We modify the DOM a bit less and loop over elements far less often. Now,
each float banner is managed by a $.fn.floatReplyDraftBanner function,
which connects window resize and scroll events directly to that banner.
This is a bit nicer than iterating over the DOM looking for float banners
to update each time, since it's far faster to get the banners we care about.

The code to create a draft banner has also been pulled out into a
$.replyDraftBanner function.