Show recommendations for certain selections during rb-site installation.
We list sqlite and mod_python as valid options during site creation, but
these are not recommended. sqlite is not for production use, and
mod_python is deprecated.

We now list text after the option indicating if the option is recommended
(in the case of wsgi), not supported for production use (sqlite), or
not supported in general (mod_python).