Fix saving comments on the review form.
The review form was broken, due to the move to the new web API in the
web UI. We were using the old style of creating datastore functions, and
ended up not having valid IDs to pass. Now we use the proper create*
and ready() functions to make sure we have valid objects we can use.

Some of this is a bit request-heavy now, and a bit ugly. I'm planning a rewrite
of this form before 1.6 that will make life much nicer and greatly improve
the experience with large reviews.