Add local_site_name parameters to most sshutils functions.
sshutils now knows how to factor in LocalSites when figuring out which
ssh configuration and keys to use. A LocalSite's configuration lives in
~/.ssh/<local_site_name>/.

As part of this, I've better cleaned up the path determination for getting
the .ssh directory. In some cases, we assumed ".ssh", and in others we checked
for "ssh". We also were using Unix-style paths everywhere. Now we check
more carefully and use os.path.join.

This is still largely untested.