Prepare SCMTools for LocalSite-based SSH keys.
Some of the SCMTool classmethods gained local_site_name parameters, which
will later be used for finding the appropriate SSH keys and such.

The forms and webapi code provide the necessary local_site_name to the
functions, ensuring that the SCMTools will have the correct information
when verifying repositories.

SCMTool gained a popen() function that wraps subprocess.Popen, setting a few
common parameters and passing in a new RB_LOCAL_SITE environment variable
that rbssh will be able to use to access the correct LocalSite info.