Add per-LocalSite SSH key support to Subversion.
Subversion now supports per-LocalSite SSH keys. As part of this, the
pysvn.Client instantiation has been moved to one place, and we've consolidated
some code as a result.

Subversion has no clear way of passing extra data down to rbssh programatically,
so what we must do is generate a .subversion/<local_site_name> for every
LocalSite we need to access. When creating the client, we look for this
directory. If not found, we create it and populate it with a config file
containing the path to rbssh, with a flag to specify the local site name.

This method is unfortunately a bit hacky, but should work, and is managed
entirely by SVNTool.