Centralize the SSH testing and make it more automatic.
SSH repository testing previously required a lot of setup, in terms of
creating repositories, configuring the access to them in settings_local.py,
and configuring SSH keys.

Now we use bundled repositories, and skip the tests if the user's public key
isn't in their ~/.ssh/authorized_keys file. We need the authorized key to
run the tests.

It's assumed if that file exists and the key is in it, SSH is configured on
the system. Otherwise, the tests will fail.