Add per-LocalSite SSH keys to Bazaar.
This adds per-LocalSite SSH key suppor to Bazaar. It required some hacks
in order to feed the information through to the SSHVendor class for rbssh.
What we have to do is override a couple of transport classes (which bzrlib
does, fortunately, make easy) and then encode the local site name as part
of the URL on access.

We catch the URL early on and strip out this extra information, so bzrlib
doesn't need to deal with it. The RBRemoteSSHTransport, which is in charge
now of setting up the basic info for an SSH transport, will set the
LocalSite name from this URL and create a new SSHVendor built for this
LocalSite, which will be passed down to rbssh.