Fix displaying screenshot draft captions.
Screenshot draft captions weren't displayed on the review request page.  We
were only showing the main caption. We now look to see if there's a draft, and
if so, we try to pull that out.

We also update the draft's timestamp when saving a screenshot, in order to bump
the Last Modified time for the page. Without this, it'd take a shift-reload in
order to see the new caption.