Always show the search field.
Now that we have basic quick search, we always want to show the search
field. If actual full-text search is enabled, the field will have the "Search"
button, but otherwise it's just a field.

This will need some improvements. It's not immediately obvious what the
field is for, and the autocomplete scrolls off the side of the screen.
These will be addressed later.