Support issues in FileAttachmentComment.
This makes FileAttachmentComment inherit from BaseComment, in order to get
the issue tracking.

There is no evolution for this, in order to prevent breakages that would
happen as a result of applying an evolution to a newly generated class in
beta 3. Instead, if any database was updated since 1.6 beta 2, it should be
evolved with 'evolve --hint --execute'.

The web API doesn't yet support issue tracking for file attachment comments.