Provide mimetype-based icons for file attachments.
When saving a file attachment, we now record the mimetype as reported by the
browser. The mimetype can then be used to return an icon representing that
file.

We have a handful of icons (all from the Tango set, which is in the public
domain). There's a mapping of mimetypes to icon types (based on Tango's)
that we use to map, say, zip files to the archive icon.