Fix some breakages with the File Attachment API access.
The File Attachment API had a missing/broken DELETE handler, and the JavaScript
calling into it was using the old names. This just provides a version of the
Screenshot handlers for this, since they're basically the same.

We also set the draft caption along with the caption, and fix setting
file attachment comment text.