Improve the look and feel of file attachments.
File attachments have a look more similar to screenshots, with an icon and
"Add Comment" buttons.

This new UI is used both when adding a new file or when loading existing
files as part of the page.

This does not hook up the comment dialog. That's coming in another change.
We may also want to revisit part of the UI for this, but for now it works.