Fix up a direct use of User.get_profile().
We don't have one good place to automatically create user profiles when the user
object is created (and we can't add a dependency on the post_save signal yet
because we still support Django 1.2), so instead we have to use
Profile.objects.get_or_create(). This change fixes one place where we could try
to access a profile that didn't exist.

Fixes bug 2111