Spew log file location with level "debug".
In production environments, the log fills up with a bunch of lines about where
the log file is (every time the web server spawns a new worker). This isn't
particularly useful, so I'm going to dial it back to only do it if
settings.DEBUG is True, and only with a level of DEBUG anyway.