Add support for Private GitHub Organizations.
A private GitHub organization requires that a username be given. However,
our "GitHub (Private)" option in the New Repository form assumed the
hosting owner. This is fine for non-organization private repos, but we really
need a username when organizations come into play.

A new option is available for private orgs. We don't simply expose a field
for a username for private GitHubs because we couldn't easily populate it
with the existing entry for current setups. It's also a bit confusing since
it's not needed if you're not using an org.

To simplify things, I've also hidden the username/password fields for all
GitHub options except for private orgs. In the private orgs case, we just hide
the password field.