Add HTTP Basic Auth support to the git raw-file method.
This is a simple change which adds support for HTTP Basic Auth to the Git
SCMTool, for use with the raw-file URL.

Tested this against a server running trac behind basic auth. Trac support
is another bag of orcas but it's not getting HTTP 401s anymore.

Reviewed at http://reviews.reviewboard.org/r/2420/