Improve the drag-and-drop UI, and add file drag-and-drop.
The old screenshot drag-and-drop UI was a bit clunky, requiring proper
aim over the drop box and shifting the page around. This was made even
more clumsy if we tried to add file drag-and-drop.

Now there's a new drop UI. When dragging a file over the browser window,
a black semi-transparent overlay will appear with two very large boxes
to drop onto, one marked "Drop Screenshot" and one marked "Drop File
Attachment." These are easy to quickly drop onto, and will begin the upload
process for the proper type.

If dropping a non-image onto the screenshot box, the error appears briefly
in the center of the overlay, before the overlay fades away.

Some additional work was done to consolidate some of the code for handling
file uploading.

This also fixes file uploading on Chrome, which was broken in 1.5.x and 1.6.

Fixes bug #1993