Improve the presentation of the Manual Updates pages.
The Manual Updates pages were very ugly-looking. The text was hard to read,
sizes were all over the place, and the content was flush with the borders of
the page.

We now have the standard Review Board header with version number, and the
account navigation bar displayed. This makes it easy to get to the
documentation and report bugs. It also adds margins to the sides.

The text size for the boxes are larger, making it easier to read and
diagnose.