Add a manual updates page for Review Board version checks.
Sometimes users get themselves into a state where things break because a site
wasn't upgraded. The best way to check for this is to compare the Review Board
version that's running with the stored version from the upgrade.

We now perform this check at startup and display an error page with
instructions on fixing the error.

Depending on whether this is (or appears to be) a developer install, we
show one set of instructions or another. Developer installs will be told to
run syncdb and evolve --execute. Production installs will be told to run
rb-site upgrade.