Add a manual updates page for database errors.
One of the most "help" pages for Review Board administrators is shown when
there's a database error. We tell them that the database just needs to be
synced/evolved, but the true cause may be that the database is corrupted,
locked, or something else.

Now we show a page saying there's a database error, recommending some
steps, and showing the actual error text to help diagnose the problem.

The old, confusing error page has been removed.