Alert the user when leaving the page while editing text.
We were alerting the user when navigating away if a comment dialog was open
and being edited, but many users were wanting this for other fields too.

So now we alert the user if any text field is being edited, if the review
form is up, or if a comment dialog is dirty.

Fixes bug #1794