Update LDAP account sync to allow for more flexibility in name attributes.
Added the following LDAP options:

* LDAP_GIVEN_NAME_ATTRIBUTE
* LDAP_SURNAME_ATTRIBUTE
* LDAP_FULL_NAME_ATTRIBUTE

While comments made in http://reviews.reviewboard.org/r/399/ are true (sn and
givenName are "standard"), they're not required or mandated to use the same
naming conventions.  They are listed as "good practice" and "acceptable" in the
related RFCs but are up to the whim of the sysadmin building the system.
Because of this, having flexibility is beneficial.  None of these fields are
required and the Full Name Attribute will always be ignored if a response is
given for Given Name or Surname.

Tested against LDAP dev LDAP server with long names (greater than two) and
changing name attributes for first_name and last_name.

Fixes bug 2122
Reviewed at http://reviews.reviewboard.org/r/2389/