Provide a script (and run it on upgrade) to fix/reset review request counters.
It's possible for review request counters in the profile and in groups to be
incorrect, especially when the system was using an older alpha/beta of 1.6.
We now have a management script for resetting these counters. It's run
automatically on upgrade, which will cover any users upgrading from an
earlier 1.6, as well as any future users if we fix any bugs in our counters
code.