Revert the addition of the diff_revision field.
The diff_revision field, while useful, caused upgrades to take forever.
Our deployment at reviews.reviewboard.org took nearly 15 minutes, and it's
a fairly small deployment. At large companies, it could take a day.

The problem is that the evolution required that every single DiffSet table be
updated, and there can easily be millions. This is impractical. What we'll
need to do is work on a new solution that can lazily migrate over, or fall
back on the old field if not populated.

The existing evolution remains in order to not break the evolution history,
but the raw SQL has been removed so that it won't execute. A new evolution
has been added that removes the field again.