Fix close descriptions on older review requests.
The logic for determining the current close description text broke on
older review requests. It assumed that the latest change description was
a status change, but this wouldn't be true for any older ones. It would
happily use the text, which was confusing.

Now the view checks specifically for a proper status change, and passes
the appropriate text as a "close_description" parameter to the template.
The template no longer has to know about the latest change description
when figuring out the text.