Polish the user page.
This adds some polish to the user page to make it a bit more clean.

First, the title now lists the user's name instead of "My Page."

Second, the sidebar has been organized to be less wordy, and to look closer
to the user info hover bubble. The labels are gone (except for "Groups"),
the sizes have changed, and the links are blue and underlined.

The e-mail address isn't shown anymore if the user isn't logged in. Same
with the info bubble.

And finally, there's a style fix for the info bubble, where the "Joined"
text was the wrong size.