Tweak the styles of the issue bars.
This just updates the icons and styles a bit for the "An issue was opened"
bars. The buttons are now system buttons again, for consistency. The
"Open issue" icon is now a red circle instead of a warning icon. The
"Issue was dropped" is a gray circle with an "X".