Fix a JavaScript error when canceling a newly created, empty reply.
Canceling a newly created but empty reply to a comment attempted to delete
the comment from the server. However, being new, there was nothing to
delete, and we didn't even have any link handy for deleting. This caused
an invalid variable reference. Now we just bail if we don't have an ID.