Fix the error text on Chrome when navigating away from a page while editing.
On Chrome, opening a field to edit and then navigating away or closing
Chrome would result in an error dialog with the text "true". This was due to
a difference in how Chrome handles the onbeforeunload event.

Like Firefox, it requires returning a string for the text. To handle IE
(which requires setting the text in event.returnValue), we would set the text
in returnValue and then return that string. Chrome, however, ignores any
attempt to set this text, so it remains "true".

Now we create the message, store it in a variable, attempt to set on
returnValue, and then just return the string variable itself, fixing all
browsers.

Fixes bug #2178