Fix closing issues bumping the review timestamp.
Closing open issues would bump the review timestamp, causing the reviews
to reorder confusingly. The logic for bumping the review timestamp made
the assumption that the review was not public (because the only time you'd
save a comment, previously, is when it was in draft form).

Now we only bump the review timestamp if we know it's not public. This
code has been moved to BaseComment so we only have to do it once, and it
also bumps the comment's own timestamp (which used to be modified as a
side-effect of saving the review).