Use a .ico favicon that works with Internet Explorer
The .png file used for the default favicon does not show up in Internet Explorer
9 (neither in the address bar nor on bookmarks).  I took the .svg-formatted
ReviewBoard logo and used Inkscape and IcoFX to create a .ico file containing
versions of the ReviewBoard logo at various sizes (16x16, 24x24, 32x32, 48x48).
This image can be used in place of the existing favicon.png and should enable
Internet Explorer to display the ReviewBoard favicon in the address bar as well
as in bookmarks and on the taskbar (when ReviewBoard is pinned to the taskbar in
Windows 7).

Testing done: These changes were tested on my local installation of 1.6 beta 2
(not re-built from source).  I personally validated the changes in IE9 and
Firefox 4.0.1, and a colleague reported that Chrome handled the changes as well.

Reviewed at http://reviews.reviewboard.org/r/2407/